Add: add api stat for account and branch

+++ src/main/java/com/example/coffee_shop_chain_management/service/AccountService.java
@@ -9,8 +9,10 @@ import com.example.coffee_shop_chain_management.repository.AccountRepository;
 import com.example.coffee_shop_chain_management.repository.BranchRepository;
 import com.example.coffee_shop_chain_management.response.APIResponse;
 import com.example.coffee_shop_chain_management.response.AccountResponse;
+import com.example.coffee_shop_chain_management.response.AccountStatResponse;
 import com.example.coffee_shop_chain_management.response.BranchResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -29,6 +31,8 @@ public class AccountService {
     private PasswordEncoder passwordEncoder;
     @Autowired
     private SendOTP sendOTP;
+    @Value("${telegram.bot.link}")
+    private String TELEGRAM_BOT_LINK;
 
     public APIResponse<List<AccountResponse>> getAllAccounts(){
         List<Account> accounts = accountRepository.findAll();
@@ -42,18 +46,28 @@ public class AccountService {
             return new APIResponse<>(null, "Account already exists!", false);
         }
 
+        Branch branch =  branchRepository.findById(accountDTO.getBranchID()).get();
+        // Neu co tai khoan co branch da ton tai thi khong tao tai khoan moi
+        if (accountRepository.existsByBranch_BranchID(branch.getBranchID())) {
+            return new APIResponse<>(null, "This branch has another account! ", false);
+        }
+
         Account account = new Account();
         account.setUsername(accountDTO.getUsername());
         account.setPassword(passwordEncoder.encode("1234"));
+        account.setRole("MANAGER");
         StringBuilder message = new StringBuilder();
         message.append("Your account is ").append("\n");
         message.append("Username: ").append(accountDTO.getUsername()).append("\n");
-        message.append("Password: ").append("1234");
-        sendOTP.sendMail(message.toString() ,accountDTO.getEmail());
-        account.setEmail(accountDTO.getEmail());
-        account.setRole("MANAGER");
+        message.append("Password: ").append("1234").append("\n");
+        message.append("Click here to verify with telegram: ").append(TELEGRAM_BOT_LINK).append(accountDTO.getUsername());
+        boolean sendStatus = sendOTP.sendMail(message.toString() ,accountDTO.getEmail());
 
-        Branch branch =  branchRepository.findById(accountDTO.getBranchID()).get();
+        if (!sendStatus) {
+            return new APIResponse<>(null, "Send email failed!", false);
+        }
+
+        account.setEmail(accountDTO.getEmail());
         account.setBranch(branch);
 
         Account newAccount = accountRepository.save(account);
@@ -129,6 +143,10 @@ public class AccountService {
             return new APIResponse<>(null, "Account not found!", false);
         }
 
+        if (account.getRole().equals("ADMIN")) {
+            return new APIResponse<>(null, "Cannot delete admin account!", false);
+        }
+
         accountRepository.delete(account);
         return new APIResponse<>(null, "Account deleted successfully!", true);
     }
@@ -139,8 +157,19 @@ public class AccountService {
             return new APIResponse<>(null, "Account not found!", false);
         }
 
+        if (accountRepository.findById(id).get().getRole().equals("ADMIN")) {
+            return new APIResponse<>(null, "Cannot delete admin account!", false);
+        }
+
         accountRepository.deleteById(id);
         return new APIResponse<>(null, "Account deleted successfully!", true);
+    }
+
+    public APIResponse<AccountStatResponse> getAccountStat(){
+        List<Account> accounts = accountRepository.findAll();
+        int totalAccounts = accounts.size();
+
+        return new APIResponse<>(new AccountStatResponse(totalAccounts), "Account statistics retrieved successfully!", true);
     }
 
     public AccountResponse toAccountResponse(Account account){
